[C64] IDE64 WIP. (nw)
git-svn-id: svn://messdev.no-ip.org/mess@15035 9849079f-6229-0410-a07f-84406ce305ba

+++ src/mess/machine/c64_ide64.h
@@ -55,7 +55,7 @@ private:
 	required_device<device_t> m_ide;
 
 	UINT8 m_bank;
-	UINT8 m_ide_data;
+	UINT16 m_ide_data;
 	int m_wp;
 	int m_enable;
 	int m_rtc_ce;
